requirements.txt
attrs==22.1.0
click==8.1.3
coverage==6.4.4
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
iniconfig==1.1.1
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
PyJWT==2.5.0
pyparsing==3.0.9
pytest==7.1.3
pytest-cov==3.0.0
python-dotenv==0.21.0
six==1.16.0
SQLAlchemy==1.4.41
tomli==2.0.1
Werkzeug==2.2.2
